# rpc: faster getblockstats using BlockUndo data