--- a/doc/release-notes-14802.md
+++ b/doc/release-notes-14802.md
@@ -1,0 +1,3 @@
+RPC changes
+-----------
+The `getblockstats` RPC is faster for fee calculation by using BlockUndo data. Also, `-txindex` is no longer required and `getblockstats` works for all non-pruned blocks.