… test/functional/data/rpc_getblockstats.json