@gmaxwell 2018-11-25
<3  I like this direction, certainly having a RPC that requires txindex is unfortunate, as does having rpcs that don't work (well) with pruning.  I don't think it would have been worth it to store extra data to make stats fast, but since we already are...