@laanwj 2019-01-22
Good to remove reliance on the transaction index.
utACK 602dd542a80d76225285ad9d52ebfc2a4d4e7c54