@MarcoFalke 2019-05-03
@FelixWeis You don't actually need to rebase. If you address my style-nits, the conflict will solve itself.