@MarcoFalke 2019-05-10
re-utACK d20d7567528e216badb8475df298bb3cec008985

Only changes are:
* Make tests more deterministic
* Add and use GetUndoChecked
* Minor stylistic changes

<details><summary>Show signature and timestamp</summary>

Signature:

```
-----BEGIN PGP SIGNED MESSAGE-----
Hash: SHA512

re-utACK d20d7567528e216badb8475df298bb3cec008985

Only changes are:
* Make tests more deterministic
* Add and use GetUndoChecked
* Minor stylistic changes
-----BEGIN PGP SIGNATURE-----

iQGzBAEBCgAdFiEE+rVPoUahrI9sLGYTzit1aX5ppUgFAlwqrYAACgkQzit1aX5p
pUh/4QwAwLZrUZJqLo1ctjMWO2SQ+ut4rxMDYaMNSIGOxw61u8J6OIUPsVfKBJLF
s2Gi0Q7XwCwlttNVDARBCQFH4R8IhVXLaY//BCfVPamVwXCO3s1x7GDtLZZrjs2c
Rng9VOhCSZ7wXZmgoQmYwj77pQOU2k1B6FXkMylzf/pt+W/jaANCuGtC4k4k8x/l
sM2sH75nWs2URra42exUbmDoIfohOHWfAxcA0Ni31AWkeRYRazWGmsvA93qL6PtP
/wN5rMNi10gCI83/N16ZZ+iq7djqRoCjkbkqSNd2D0NJOmquEIN4W3GGQJ1BciGF
nu0gLQ12ShXCVf167E6s4xGPsKu9hsiUNJN9xursiD1RUhGtUQXdpIRjodL+lDj8
CZhHeUj0whAXKkS+YddhPO385rzqChebvdWZL5stRlUi4TFKzJM+rvsjGNaEKgxO
wjiUfi1L6v07zstlKFLc1pfri9Yws5b4MT6fO9+9dQrtUm1IcS4NeS4YrKKaiI4X
B9R6OHC3
=xoBb
-----END PGP SIGNATURE-----
```

[Timestamp of file with hash `5dc2003a827858f9302ac84ff924bdf15ada47e85278dce1eee93d098c8e8abd  -`](https://opentimestamps.org/info/?004f70656e54696d657374616d7073000050726f6f6600bf89e2e884e8929401085dc2003a827858f9302ac84ff924bdf15ada47e85278dce1eee93d098c8e8abdf01001b31b759d70096887582f20c060390608fff0103a16fd4f7c8f24396601d6d3df4b77c308f1045cd5b2aff008685132d3134760540083dfe30d2ef90c8e232268747470733a2f2f6274632e63616c656e6461722e636174616c6c6178792e636f6dfff0107a163f42ff2271de93438f4e185e505108f1045cd5b2aff008b2ee9ad2f7ef444d0083dfe30d2ef90c8e2e2d68747470733a2f2f616c6963652e6274632e63616c656e6461722e6f70656e74696d657374616d70732e6f7267fff01096e0a180f06409fa2b145fd3964ed5f808f1045cd5b2aff0084b4d68869d984add0083dfe30d2ef90c8e2c2b68747470733a2f2f626f622e6274632e63616c656e6461722e6f70656e74696d657374616d70732e6f7267f010ae41a855e90c95af1d865a95e775f1a708f010e234702407890bd5586ea5cfbfa091cb08f1045cd5b2aff008ce6559de452581c00083dfe30d2ef90c8e292868747470733a2f2f66696e6e65792e63616c656e6461722e657465726e69747977616c6c2e636f6d)

</details>